code cleanup and move Enums to folder
