+++ ThreeXPlusOne.UnitTests/DirectedGraphTests.cs
@@ -3,9 +3,9 @@ using Microsoft.Extensions.Options;
 using Moq;
 using System.Collections.ObjectModel;
 using ThreeXPlusOne.Code.Graph;
-using ThreeXPlusOne.Code.Graph.Services;
 using ThreeXPlusOne.Code.Interfaces;
 using ThreeXPlusOne.Config;
+using ThreeXPlusOne.Enums;
 using ThreeXPlusOne.Models;
 using ThreeXPlusOne.UnitTests.Mocks;
 using Xunit;
