+++ ThreeXPlusOne/Models/Shape.cs
@@ -1,4 +1,5 @@
 using System.Drawing;
+using ThreeXPlusOne.Enums;
 
 namespace ThreeXPlusOne.Models;
